requirements.txt
absl-py
ipykernel
numpy
pandas
scipy
requests
statsmodels
pytest
biopython
